Add scripts for testing and fix bugs
In the last version, when posting job to the server, the server will set
a timer to retrieve the job status and its solution when it times out.
However, the db only update the computation time instead of the status.
The status remains in the last status. In this commit, fixed.