replace getStaticProps with getServerSideProps
In the last commit, the data in the history page would be cached since
the getStaticProps is used to fetch the data as the page is requested
firstly. The data will be cached so that when reload the page, the data
reamin the same. In this commit, the function is replaced with the
get ServerSideProps.